dom/workers/test/test_bug949946.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 949946</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  SpecialPowers.pushPrefEnv({ set: [["dom.workers.sharedWorkers.enabled", true]] }, function() {
    new SharedWorker('sharedWorker_sharedWorker.js');
    new SharedWorker('sharedWorker_sharedWorker.js', ':');
    new SharedWorker('sharedWorker_sharedWorker.js', '|||');
    ok(true, "3 SharedWorkers created!");
    SimpleTest.finish();
  });

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
